Move controller tests to different file
